+++ packages/go_router/lib/src/route_data.dart
@@ -59,7 +59,7 @@ abstract class GoRouteData {
   /// A helper function used by generated code.
   ///
   /// Should not be used directly.
-  static String $location(String path, {Map<String, String>? queryParams}) =>
+  static String $location(String path, {Map<String, dynamic>? queryParams}) =>
       Uri.parse(path)
           .replace(
             queryParameters: