+++ packages/pigeon/tool/shared/test_suites.dart
@@ -216,18 +216,6 @@ Future<int> _runFlutterUnitTests() async {
     'non_null_fields',
     'null_fields',
     'nullable_returns',
-    // TODO(stuartmorgan): Eliminate these files by ensuring that everything
-    // they are intended to cover is in core_tests.dart (or, if necessary in
-    // the short term due to limitations in non-Dart generators, a single other
-    // file). They aren't being unit tested, only analyzed.
-    'async_handlers',
-    'host2flutter',
-    'list',
-    'message',
-    'void_arg_flutter',
-    'void_arg_host',
-    'voidflutter',
-    'voidhost',
   ];
   final int generateCode = await _generateDart(<String, String>{
     for (final String name in inputPigeons)
@@ -315,7 +303,7 @@ Future<int> _runIOSPluginUnitTests(String testPluginPath) async {
   return runXcodeBuild(
     '$examplePath/ios',
     sdk: 'iphonesimulator',
-    destination: 'platform=iOS Simulator,name=iPhone 8',
+    destination: 'platform=iOS Simulator,name=iPhone 13',
     extraArguments: <String>['test'],
   );
 }